Added ResponseContainer moved some object-classes
