+++ src/main/java/objects/response_container/ResponseContainer.java
@@ -1,0 +1,11 @@
+package objects.response_container;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ResponseContainer<T> {
+    private int statusCode;
+    private T data;
+}
