+++ src/main/java/objects/support/Support.java
@@ -1,4 +1,4 @@
-package objects;
+package objects.support;
 
 import lombok.Builder;
 import lombok.Data;
